database/requirements.txt
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
python-dotenv==1.1.1
tqdm==4.67.1
typing-extensions==4.15.0
